[#37] refactor: import 경로 변경
