+++ client/recoil/addressState.ts
@@ -1,12 +1,10 @@
 import { atom } from "recoil";
-import SignUpType from "@pages/signup/SignUpType";
+import SignUpType from "types/SignUpType";
 
-export const addressState = atom<Pick<SignUpType, "address" | "addressDetail">>(
-	{
-		key: "address",
-		default: {
-			address: "",
-			addressDetail: "",
-		},
-	}
-);
+export const addressState = atom<Pick<SignUpType, "address" | "addressDetail">>({
+	key: "address",
+	default: {
+		address: "",
+		addressDetail: "",
+	},
+});
